cell_model_passports_models is a preprocessed version of the cell model passports dataset

A preprocessed version of the cell model passports dataset. This dataset contains the following columns: CMP.model_id, CMP.sampleid, CMP.model_name, CMP.cancer_type_ncit_id, CMP.COSMIC_ID, CMP.BROAD_ID, CMP.CCLE_ID, and CMP.RRID. This dataset is used in the AnnotationGx package to map cell line names from various sources to the Cellosaurus database.

Usage

data(cell_model_passports_models)

Format

A data table with 8 columns and 1001 rows.

CMP.model_id

char The model id of the cell line in the cell model passports dataset.

CMP.sample_id

char The sample id of the cell line in the cell model passports dataset.

CMP.model_name

char The name of the cell line in the cell model passports dataset.

CMP.cancer_type_ncit_id

char The cancer type ncit id of the cell line in the cell model passports dataset.

CMP.COSMIC_ID

int The COSMIC ID of the cell line in the cell model passports dataset.

CMP.BROAD_ID

int The BROAD ID of the cell line in the cell model passports dataset.

CMP.CCLE_ID

int The CCLE ID of the cell line in the cell model passports dataset.

CMP.RRID

char The RRID of the cell line in the cell model passports dataset.

Source

https://cog.sanger.ac.uk/cmp/download/model_list_20240103.csv

Examples

data(cell_model_passports_models)
head(cell_model_passports_models)
#>    CMP.model_id CMP.sample_id CMP.model_name CMP.cancer_type_ncit_id
#>          <char>        <char>         <char>                  <char>
#> 1:    SIDM01774     SIDS01659          PK-59                   C3850
#> 2:    SIDM00192     SIDS00612       SNU-1033                   C9383
#> 3:    SIDM01447     SIDS01466        SNU-466                   C3058
#> 4:    SIDM01554     SIDS01363      IST-MES-2                  C45662
#> 5:    SIDM01689     SIDS01631         MUTZ-5                   C8644
#> 6:    SIDM01460     SIDS01493          TM-31                  C60781
#>    CMP.COSMIC_ID CMP.BROAD_ID                              CMP.CCLE_ID
#>           <char>       <char>                                   <char>
#> 1:                 ACH-000205                            PK59_PANCREAS
#> 2:                 ACH-000286                  SNU1033_LARGE_INTESTINE
#> 3:                 ACH-000289            SNU466_CENTRAL_NERVOUS_SYSTEM
#> 4:                 ACH-000331                           ISTMES2_PLEURA
#> 5:                 ACH-000492 MUTZ5_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE
#> 6:                 ACH-000592              TM31_CENTRAL_NERVOUS_SYSTEM
#>     CMP.RRID
#>       <char>
#> 1: CVCL_4897
#> 2: CVCL_5002
#> 3: CVCL_5064
#> 4: CVCL_1312
#> 5: CVCL_1873
#> 6: CVCL_6735