A general function for creating Queries to the ChEMBL API

A general function for creating Queries to the ChEMBL API

Resource Name	Description

activity	Activity values recorded in an Assay

assay	Assay details as reported in source Document/Dataset

atc_class	WHO ATC Classification for drugs

binding_site	WHO ATC Classification for drugs

biotherapeutic	Biotherapeutic molecules, which includes HELM notation and sequence data

cell_line	Cell line information

chembl_id_lookup	Look up ChEMBL Id entity type

compound_record	Occurence of a given compound in a spcecific document

compound_structural_alert	Indicates certain anomaly in compound structure

document	Document/Dataset from which Assays have been extracted

document_similarity	Provides documents similar to a given one

document_term	Provides keywords extracted from a document using the TextRank algorithm

drug	Approved drugs information, including (but not limited to) applicants, patent numbers and research codes. This endpoint aggregates data on the parent, please use the parent chembl id found in other endpoints

drug_indication	Joins drugs with diseases providing references to relevant sources

drug_warning	Safety information for drugs withdrawn from one or more regions of the world and drugs that carry a warning for severe or life threatening adverse effects

go_slim	GO slim ontology

image	Graphical (svg) representation of Molecule

mechanism	Mechanism of action information for approved drugs

metabolism	Metabolic pathways with references

molecule	Molecule information, including properties, structural representations and synonyms

molecule_form	Relationships between molecule parents and salts

organism	Simple organism classification

protein_classification	Protein family classification of TargetComponents

similarity	Molecule similarity search

source	Document/Dataset source

status	API status with ChEMBL DB version number and API software version number

substructure	Molecule substructure search

target	Targets (protein and non-protein) defined in Assay

target_component	Target sequence information (A Target may have 1 or more sequences)

target_relation	Describes relations between targets

tissue	Tissue classification

xref_source	Cross references to other resources for compounds

Filter Type	Description

exact (iexact)	Exact match with query (case insensitive equivalent)

contains (icontains)	Wild card search with query (case insensitive equivalent)

startswith (istartswith)	Starts with query (case insensitive equivalent)

endswith (iendswith)	Ends with query (case insensitive equivalent)

regex (iregex)	Regular expression query (case insensitive equivalent)

gt (gte)	Greater than (or equal)

lt (lte)	Less than (or equal)

range	Within a range of values

in	Appears within list of query values

isnull	Field is null

search	Special type of filter allowing a full text search based on elastic search queries

only	Select specific properties from the original endpoint and returns only the desired properties on each record

Usage

constructChemblQuery(resource, field, filter_type, value)