Upload and optionally publish data to Zenodo

Make the PUT and POST request necessary to upload and optionally publish a file to Zenodo.

Usage

depositZenodo(
  file_path,
  metadata = zenodoMetadata(),
  publish = FALSE,
  ...,
  url = "https://zenodo.org/api/deposit/depositions",
  access_token = Sys.getenv("ZENODO_TOKEN")
)

Arguments

file_path

character(1) Path to the file to upload to Zenodo.

metadata

list Named list of metadata required to publish publish a Zenodo entry, as returned by the zenodoMetadata function. See ?zenodoMetadata for documentation of the correct format.

publish

logical(1) Should the Zenodo entry be published, assuming the upload succeeds? Default is FALSE.

...

Fall through parameters to the httr::POST and httr::PUT method calls inside this function. Pass httr::verbose() to get detailed information about your API requests, useful for debugging.

url

character(1) URL string for the API. Do not change unless you know what you are doing.

access_token

character(1) Zenodo access token. By deafult tries to read this from the ZENODO_TOKEN environmet variable via Sys.getenv. See details for more information.

Value

If all HTTP requests succeed, returns the details of the Zenodo entry as a list invisiblty. If any of the requests fail, it returns the failed request object for ease of debugging.

Details

Caveats

This function is an early prototype and is not very polished. The current version will create a new entry everytime the function is called, even if some of the steps succeeded. For now, you will need to manually delete these entries from you Zenodo account. However, this will be corrected in future updates.

access_token

API documentation is available at https://developers.zenodo.org/.

To use this function you need a Zenodo access token for your account. A link is available in the API documentation to create one. It is assumed you have added the access token to your operating system environment in the ZENODO_TOKEN variable.

If you have not, you can specify an access token. For this function to work the token needs to provide both upload and publish permissions on your Zenodo account.

Trouble Shooting

405 Errors: Ensure that you have verified your Zenodo account email.

403 Errors: Ensure you have checked both the publish:actions and publish:write boxes for your Zenodo access token.

See also