Build a query for the PubChem REST API

This function builds a query for the PubChem REST API based on the provided parameters.

Usage

.build_pubchem_rest_query(
  id,
  domain = "compound",
  namespace = "name",
  operation = "cids",
  output = "JSON",
  url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug",
  raw = FALSE,
  query_only = FALSE,
  ...
)

Arguments

id

The identifier(s) for the query. If namespace is 'name', id must be a single value.

domain

The domain of the query. Options are 'compound', 'substance', 'assay', 'cell', 'gene', 'protein'.

namespace

The namespace of the query. Options depend on the chosen domain.

operation

The operation to perform. Options depend on the chosen domain and namespace.

output

The desired output format. Options are 'JSON', 'XML', 'SDF', 'TXT', 'CSV'.

url

The base URL for the PubChem REST API.

raw

Logical indicating whether to return the raw response or parse it.

query_only

Logical indicating whether to return the query URL only.

...

Additional arguments to be passed to the query.

Value

The query URL or the parsed response, depending on the arguments.