Query the ChEMBL mechanism API end-point for a given CHEMBL_ID or list of CHEMBL_IDs

A specialised function for querying the mechanism resource from the ChEMBL API.

Usage

getChemblMechanism(
  chembl.ID,
  resources = "mechanism",
  field = "molecule_chembl_id",
  filter_type = "in",
  returnURL = FALSE
)

Arguments

chembl.ID

character(1) or character(n) CHEMBL_ID(s) to query

resources

character(1) Resource to query. Default is "mechanism"

field

character(1) or character(n) Field(s) to query. Default is "molecule_chembl_id"

filter_type

character(1) Filter type. Default is "in"

returnURL

logical(1) Return the URL(s) instead of the data table

Value

data.table A table containing all mechanism of action entries and the CHEMBL_IDs queried

Details

Query the ChEMBL mechanism API end-point for a given CHEMBL_ID or list of CHEMBL_IDs

Examples

getChemblMechanism("CHEMBL1413")
#>    Molecule Chembl ID     Action Type       Mechanism Of Action
#>                <char>          <char>                    <char>
#> 1:         CHEMBL1413 CHELATING AGENT      Iron chelating agent
#> 2:         CHEMBL1413 CHELATING AGENT Aluminium chelating agent
#>    Molecular Mechanism
#>                  <int>
#> 1:                   1
#> 2:                   1
#>                                                                                                                                                  Mechanism Comment
#>                                                                                                                                                             <char>
#> 1: Trivalent metal cations chelating agent; inhibition of the metal-dependent enzymes that are responsible for the degradation of peroxides within the fungal cell
#> 2: Trivalent metal cations chelating agent; inhibition of the metal-dependent enzymes that are responsible for the degradation of peroxides within the fungal cell
#>    Parent Molecule Chembl ID Target Chembl ID
#>                       <char>           <char>
#> 1:                CHEMBL1413    CHEMBL2363058
#> 2:                CHEMBL1413    CHEMBL2366381

chembl_ids <- c("CHEMBL515", "CHEMBL235191", "CHEMBL1413")
getChemblMechanism(chembl_ids)
#>    Molecule Chembl ID     Action Type           Mechanism Of Action
#>                <char>          <char>                        <char>
#> 1:          CHEMBL515       INHIBITOR                 DNA inhibitor
#> 2:       CHEMBL235191       INHIBITOR Histone deacetylase inhibitor
#> 3:         CHEMBL1413 CHELATING AGENT          Iron chelating agent
#> 4:         CHEMBL1413 CHELATING AGENT     Aluminium chelating agent
#>    Molecular Mechanism
#>                  <int>
#> 1:                   1
#> 2:                   1
#> 3:                   1
#> 4:                   1
#>                                                                                                                                                  Mechanism Comment
#>                                                                                                                                                             <char>
#> 1:                                                                                                                                                Alkylating agent
#> 2:                                                                                                                                                            <NA>
#> 3: Trivalent metal cations chelating agent; inhibition of the metal-dependent enzymes that are responsible for the degradation of peroxides within the fungal cell
#> 4: Trivalent metal cations chelating agent; inhibition of the metal-dependent enzymes that are responsible for the degradation of peroxides within the fungal cell
#>    Parent Molecule Chembl ID Target Chembl ID
#>                       <char>           <char>
#> 1:                 CHEMBL515    CHEMBL2311221
#> 2:              CHEMBL235191    CHEMBL2093865
#> 3:                CHEMBL1413    CHEMBL2363058
#> 4:                CHEMBL1413    CHEMBL2366381