getPubChemFromNSC

Return a data.table mapping from ids to the information specified in to.

Usage

getPubChemFromNSC(
  ids,
  to = "cids",
  ...,
  batch = TRUE,
  raw = FALSE,
  proxy = FALSE,
  options = NA,
  query_only = FALSE
)

Arguments

ids

A character or numeric vector of valid NSC ids to use for the query.

to

A character(1) vector with the desired return type. Currently only 'cids' and 'sids' are implemented, but other options are available via the PubChem API. This corresponds to the operation portion of the PubChem API URL Path.

...

Fall through arguments to bpmapply. Use this to pass in BPPARAM parameter to customize parellization settings. Alternatively, just call register() with your desired parallel backend configuration.

raw

A logical(1) vector specifying whether to early return the raw query results. Use this if specifying an unimplemented return to the to parameter.

proxy

logical(1) Should the query be routed through a random proxy server. This is useful to keep trying queries if a user gets blacklisted.

Value

A data.table where the first column is the specified NSC ids and the second column is the results specified in to.