Retrieve PubChem substance information

This function retrieves PubChem substance information based on the provided IDs. It supports mapping from various namespaces such as CID, name, xref, smiles, inchi, and sdf. The function can handle batch queries and can return the raw query results if specified.

Usage

getPubChemSubstance(
  ids,
  from = "cid",
  to = "sids",
  ...,
  batch = TRUE,
  raw = FALSE
)

Arguments

ids

A character vector of IDs to query.

from

The namespace of the input IDs (default is 'cid').

to

The namespace to map the input IDs to (default is 'sids').

...

Additional arguments to be passed to the queryPubChem function.

batch

Logical indicating whether to use batch queries (default is TRUE).

raw

Logical indicating whether to return the raw query results (default is FALSE).

Value

A data table containing the retrieved PubChem substance information. If raw is TRUE, the function returns the raw query results. If any queries fail, the function includes the failed queries in the returned object.

References

Documentation for the PubChem API: https://pubchemdocs.ncbi.nlm.nih.gov/

Examples

getPubChemSubstance(c('123', '456'), from='cid', to='sids')
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> Null data.table (0 rows and 0 cols)
getPubChemSubstance(c('water', 'ethanol'), from='name', to='cid', batch=FALSE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
#> 
#> Null data.table (0 rows and 0 cols)
getPubChemSubstance(c('C1=CC=CC=C1', 'CCO'), from='smiles', to='name', raw=TRUE)
#> Warning: Batch queries cannot be used when mapping from name, xref, smiles,
#> inchi or sdf. Setting to batch=FALSE.
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
#> 
#> [[1]]
#> [[1]]$Fault
#> [[1]]$Fault$Code
#> [1] "PUGREST.BadRequest"
#> 
#> [[1]]$Fault$Message
#> [1] "Invalid identifier namespace"
#> 
#> [[1]]$Fault$Details
#> [1] "Expected identifier namespace: sid, sourceid, sourceall, name, xref, or listkey"
#> 
#> 
#> 
#> [[2]]
#> [[2]]$Fault
#> [[2]]$Fault$Code
#> [1] "PUGREST.BadRequest"
#> 
#> [[2]]$Fault$Message
#> [1] "Invalid identifier namespace"
#> 
#> [[2]]$Fault$Details
#> [1] "Expected identifier namespace: sid, sourceid, sourceall, name, xref, or listkey"
#> 
#> 
#>