Constructs and executes a GET request to the PubChem PUG REST API

This function builds a query URL for the PubChem PUG REST API based on the function parameters then executes that query, returning a httr::request object.

Most – if not all – of the information PubChem PUG service needs to produce its results is encoded into the URL. The general form of the URL has three parts – input, operation, and output – after the common prefix, followed by operation options as URL arguments (after the ‘?’):

"https://pubchem.ncbi.nlm.nih.gov/rest/pug///[][?<operation_options>]"

Input

The input portion of the URL tells the service which records to use as the subject of the query. This is further subdivided into two or more locations in the URL “path” as follows:

" = //"

" = substance | compound | assay | "

"compound domain = cid | name | smiles | inchi | sdf | inchikey | formula | | | listkey | "

"substance domain = sid | sourceid/ | sourceall/ | name | | listkey"

"assay domain = aid | listkey | type/ | sourceall/ | target/ | activity/"

Complete documentation for valid input specifications can be found at: https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865556

Operation

The operation part of the URL tells the service what to do with the input records – such as to retrieve whole record data blobs or specific properties of a compound, etc. The construction of this part of the “path” will depend on what the operation is. Currently, if no operation is specified at all, the default is to retrieve the entire record. What operations are available are, of course, dependent on the input domain – that is, certain operations are applicable only to compounds and not assays.

"compound domain = record | | synonyms | sids | cids | aids | assaysummary | classification | | description | conformers"

"substance domain = record | synonyms | sids | cids | aids | assaysummary | classification | | description"

"assay domain = record | concise | aids | sids | cids | description | targets/ | | summary | classification"

Complete documentation for valid operations can be found at: https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865557

Output

The final portion of the URL tells the service what output format is desired. Note that this is formally optional, as output format can also be specified in the HTTP Accept field of the request header – see below for more detail.

" = XML | ASNT | ASNB | JSON | JSONP | SDF | CSV | PNG | TXT"

Complete documentation for valid output formats can be found at: https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865558

Usage

getRequestPubChem(
  id,
  domain = "compound",
  namespace = "cid",
  operation = NA,
  output = "JSON",
  ...,
  url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug",
  operation_options = NA,
  raw = FALSE,
  query_only = FALSE,
  verbose = FALSE
)

Arguments

id

The identifier for REST API query. This should be a valid compound identifier for the specified domain and namespace. For example, if domain='compound' and namespace='cid' then id should be one or more compound cids to retrieve information about. If id is a list or vector, the ids are parsed to a string with ids separated by commas. This parameter forms the '' part of the Input portion of the URL path: '//'.

domain

Which records in PubChem are the subject of the query? Default is 'compound'. Usually, this will be one of 'substance', 'compound', or 'assay'. For more advanced options see the PubChem PUG REST API documentation. This forms the '' part of the Input portion of the URL path.

namespace

What kind of identifiers are in id? These options are specific to the selected domain. For the 'compound' domain, common options are 'cid', 'name', 'smiles', 'inchi', or 'formula'. For the 'substance' domain common options are 'sid' or 'name'. For the 'assay' domain common options are 'aid' and 'target/<assay_target>' where '<assay_target>' is one of 'gi', 'proteinname', 'geneid', 'genesymbol' or 'accession'. For all options please see the PubChem PUG REST API documentation. This parameter makes up the '' part of the Input portion of the URL path.

operation

What kind of data to return for the specified input portion of the URL path? If excluded this defaults to return the entire record for that specified input. Options are domain specific and details are available in the PubChem PUG REST API documentation. Options valid for 'compound', 'substance' and 'assay' domains are 'record', 'aids', 'sids', 'aids', 'description' and 'classification'. Common options for the 'compound' domain are 'assaysummary', 'conformers' and 'property/' where '' is a comma separated list property tags; valid tags can be found at https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865565. Common options for the 'substance' domain are 'assaysummary'. Common options for the 'assay' domain include 'concise', 'summary' and 'targets/' where '' is one of 'ProteinGI', 'ProteinName', 'GeneID' or 'GeneSymbol'.

output

What format should the data be returned in? Default is 'JSON'. other common options include 'XML', 'CSV' and 'TXT'. For a complete list of output format options, plase see the PubChem PUG REST API documentation.

...

Fall through arguments to httr::GET.

url

The URL of the PubChem REST API. Probably don't change this.

operation_options

Further optional arguments for the selected operation. this is specific to the selected operation. This is appended as a string after '?' at the end of the query. See https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865565 for details.

query_only

logical(1) Should this function early return only the encoded query?

Value

A httr::response object with the results of the GET request.

Details

See the PUG_VIEW website for the complete API documentation. A subset of this documentation is included below for convenience.

References

Kim S, Thiessen PA, Cheng T, Yu B, Bolton EE. An update on PUG-REST: RESTful interface for programmatic access to PubChem. Nucleic Acids Res. 2018 July 2; 46(W1):W563-570. doi:10.1093/nar/gky294.

Kim S, Thiessen PA, Bolton EE, Bryant SH. PUG-SOAP and PUG-REST: web services for programmatic access to chemical information in PubChem. Nucleic Acids Res. 2015 Jul 1; 43(W1):W605-W611. doi: 10.1093/nar/gkv396.

Kim S, Thiessen PA, Bolton EE. Programmatic Retrieval of Small Molecule Information from PubChem Using PUG-REST. In Kutchukian PS, ed. Chemical Biology Informatics and Modeling. Methods in Pharmacology and Toxicology. New York, NY: Humana Press, 2018, pp. 1-24. doi:10.1007/7653_2018_30.

See also