Map Cell Line IDs to Accession Numbers

This function maps cell line IDs to accession numbers using the Cellosaurus database.

Usage

mapCell2Accession(
  ids,
  numResults = 10000,
  from = "idsy",
  sort = "ac",
  keep_duplicates = FALSE,
  fuzzy = FALSE,
  query_only = FALSE,
  raw = FALSE,
  parsed = TRUE,
  ...
)

Arguments

ids

A character vector of cell line IDs.

numResults

The maximum number of results to return for each query. Default is 1000.

from

The type of input IDs. Possible values are "idsy" (default), "ac", "id", "sy", and "misspelling".

sort

The sorting order of the results. Possible values are "ac" (default), "id", "sy", and "misspelling".

keep_duplicates

Logical indicating whether to keep duplicate results. Default is FALSE.

fuzzy

Logical indicating whether to perform a fuzzy search. Default is FALSE.

query_only

Logical indicating whether to return only the query URLs. Default is FALSE.

raw

Logical indicating whether to return the raw HTTP responses. Default is FALSE.

parsed

Logical indicating whether to parse the response text. Default is TRUE.

...

Additional arguments to be passed to the underlying functions.

Value

A data.table containing the mapped cell line IDs and accession numbers.

Examples

mapCell2Accession(ids = c("A549", "MCF7"))
#>    cellLineName accession  query
#>          <char>    <char> <char>
#> 1:        A-549 CVCL_0023   A549
#> 2:        MCF-7 CVCL_0031   MCF7