Maps cell line names to accession numbers

This function takes a vector of cell line names and maps them to accession numbers using the Cellosaurus database. It performs a parallel request to retrieve the mapping information and returns a data table with the results.

Usage

mapCell2Accession(
  ids,
  numResults = 1000,
  from = "id",
  to = c("id", "ac"),
  prioritizeParent = FALSE,
  query_only = FALSE,
  raw = FALSE,
  BPPARAM = BiocParallel::SerialParam(),
  ...
)

Arguments

ids

A character vector of cell line names.

numResults

The number of results to return for each query. Default is 1.

from

The field to query from. Default is "id".

to

The field to query to. Default is both "id" and "ac".

prioritizeParent

If TRUE, prioritizes the parent cell line when multiple matches are found. Default is FALSE. When prioritizeParent is TRUE, if multiple matches are found for a cell line name, the function will prioritize the parent cell lines over other matches. This can be useful when dealing with cell line hierarchies where the parent cell line represents a broader category. An example of this is trying to map using id "BT474" which returns "CVCL_YX79" which corresponds to "BT474 A3" whereas "BT-474" exists in the database as "CVCL_0179". If prioritizeParent is TRUE, the function will prioritize "CVCL_0179" over "CVCL_YX79" since "BT-474" is the parent cell line of "BT474 A3".

query_only

If TRUE, returns the query URL instead of the results. Default is FALSE.

raw

If TRUE, returns the raw response instead of a data table. Default is FALSE.

BPPARAM

A BiocParallel parameter object controlling the parallelization.

...

Additional arguments to pass to the request.

Value

Depending on parameters, either a: data.table with the "id", "ac", "query", and "query:id" columns. list of query URLs if query_only is TRUE. list of raw responses if raw is TRUE.

Examples

mapCell2Accession(c("A549", "HeLa"))
#>                         id        ac   query query:id
#>                     <char>    <char>  <char>   <char>
#>    1:            A549 Cas9 CVCL_DX70 id:A549     A549
#>    2:        A549 DHODH-/- CVCL_YZ46 id:A549     A549
#>    3:     A549 HIF1A -/-/- CVCL_C3C5 id:A549     A549
#>    4:            A549 MX10 CVCL_LB65 id:A549     A549
#>    5:             A549 OS8 CVCL_XY99 id:A549     A549
#>   ---                                                
#> 1120:  Abcam HeLa ROCK2 KO CVCL_B2E1 id:HeLa     HeLa
#> 1121:   Abcam HeLa RORC KO CVCL_B2E2 id:HeLa     HeLa
#> 1122: Abcam HeLa RPL36A KO CVCL_B2E3 id:HeLa     HeLa
#> 1123:   Abcam HeLa RPRM KO CVCL_B2E4 id:HeLa     HeLa
#> 1124: Abcam HeLa RPS27L KO CVCL_B2E5 id:HeLa     HeLa