Query UniChem for a compound.

This function queries the UniChem API for a compound based on the provided parameters.

Usage

queryUnichemCompound(
  compound,
  type,
  sourceID = NA_integer_,
  request_only = FALSE,
  raw = FALSE,
  ...
)

Arguments

compound

character or integer The compound identifier to search for.

type

character The type of compound identifier to search for. Valid types are "uci", "inchi", "inchikey", and "sourceID".

sourceID

integer The source ID to search for if the type is "sourceID". Defaults to NULL.

request_only

boolean Whether to return the request only. Defaults to FALSE.

raw

boolean Whether to return the raw response. Defaults to FALSE.

...

Additional arguments.

Value

A list with the external mappings and the UniChem mappings.

Examples

queryUnichemCompound(type = "sourceID", compound = "444795", sourceID = 22)
#> $External_Mappings
#>                  compoundID             Name
#>                      <char>           <char>
#>  1:                CHEMBL38           chembl
#>  2:                 DB00755         drugbank
#>  3:                     REA              pdb
#>  4:                    2644           gtopdb
#>  5:                12014646     pubchem_dotf
#>  6:                  C00777      kegg_ligand
#>  7:                   15367            chebi
#>  8:            SAM002264647          nih_ncc
#>  9:        ZINC000012358651             zinc
#> 10:                 1934590       emolecules
#> 11:               tretinoin            atlas
#> 12:           retinoic acid            atlas
#> 13: all-trans retinoic acid            atlas
#> 14:     9-cis retinoic-acid            atlas
#> 15:              5688UTC01R           fdasrs
#> 16:         SCHEMBL19091395       surechembl
#> 17:             SCHEMBL3145       surechembl
#> 18:             PA164746900         pharmgkb
#> 19:             HMDB0001852             hmdb
#> 20:      Tretinoin(Aberela)          selleck
#> 21:                14849563  pubchem_tpharma
#> 22:                14825303  pubchem_tpharma
#> 23:                  444795          pubchem
#> 24:                60021090      nmrshiftdb2
#> 25:               LSM-42854            lincs
#> 26:               4759-48-2            actor
#> 27:                302-79-4            actor
#> 28:             187175-63-9            actor
#> 29:     MolPort-000-883-857          molport
#> 30:             J2.378.058E          nikkaji
#> 31:             J1.313.469C          nikkaji
#> 32:               J623.910B          nikkaji
#> 33:               J970.183D          nikkaji
#> 34:                 J1.518K          nikkaji
#> 35:               J494.243D          nikkaji
#> 36:               J690.379G          nikkaji
#> 37:               J646.157C          nikkaji
#> 38:               J646.158A          nikkaji
#> 39:               J528.606I          nikkaji
#> 40:                  323588        bindingdb
#> 41:                   31883        bindingdb
#> 42:           DTXSID7021239          comptox
#> 43:            LMPR01090019        lipidmaps
#> 44:                    2722      drugcentral
#> 45:              MTBLC15367     metabolights
#> 46:                    6481           brenda
#> 47:                   12679           brenda
#> 48:                   21231           brenda
#> 49:                    2354           brenda
#> 50:                    5126           brenda
#> 51:                  193381           brenda
#> 52:                  260365           brenda
#> 53:               CB6222631     chemicalbook
#> 54:               TRETINOIN         dailymed
#> 55:                  EUDYNA   clinicaltrials
#> 56:               TRETINOIN   clinicaltrials
#> 57:                  ABEREL   clinicaltrials
#> 58:                  RENOVA   clinicaltrials
#> 59:                   AVITA   clinicaltrials
#> 60:              NSC-122758   clinicaltrials
#> 61:                VESANOID   clinicaltrials
#> 62:                 RETIN A   clinicaltrials
#> 63:                 RETIN-A   clinicaltrials
#> 64:           RETINOIC ACID   clinicaltrials
#> 65:                 ALTRENO   clinicaltrials
#> 66:                  RENOVA           rxnorm
#> 67:                   AVITA           rxnorm
#> 68:   TRETINOIN MICROSPHERE           rxnorm
#> 69:                 RETIN-A           rxnorm
#> 70:               TRETINOIN           rxnorm
#> 71:                 ACTICIN           rxnorm
#> 72:                 ATRALIN           rxnorm
#> 73:                 ALTRENO           rxnorm
#> 74:                HY-14649   MedChemExpress
#> 75:                PD001430 probes_and_drugs
#>                  compoundID             Name
#>                                                    NameLong sourceID
#>                                                      <char>    <int>
#>  1:                                                  ChEMBL        1
#>  2:                                                DrugBank        2
#>  3:                         PDBe (Protein Data Bank Europe)        3
#>  4:                                   Guide to Pharmacology        4
#>  5:                  PubChem ('Drugs of the Future' subset)        5
#>  6:   KEGG (Kyoto Encyclopedia of Genes and Genomes) Ligand        6
#>  7:       ChEBI (Chemical Entities of Biological Interest).        7
#>  8:                                 NIH Clinical Collection        8
#>  9:                                                    ZINC        9
#> 10:                                              eMolecules       10
#> 11:                                   Gene Expression Atlas       12
#> 12:                                   Gene Expression Atlas       12
#> 13:                                   Gene Expression Atlas       12
#> 14:                                   Gene Expression Atlas       12
#> 15:             FDA/USP Substance Registration System (SRS)       14
#> 16:                                              SureChEMBL       15
#> 17:                                              SureChEMBL       15
#> 18:                                                PharmGKB       17
#> 19:                        Human Metabolome Database (HMDB)       18
#> 20:                                                 Selleck       20
#> 21:                       PubChem ('Thomson Pharma' subset)       21
#> 22:                       PubChem ('Thomson Pharma' subset)       21
#> 23:                                       PubChem Compounds       22
#> 24:                                              NMRShiftDB       24
#> 25: Library of Integrated Network-based Cellular Signatures       25
#> 26:                                                   ACToR       26
#> 27:                                                   ACToR       26
#> 28:                                                   ACToR       26
#> 29:                                                 MolPort       28
#> 30:                                                 Nikkaji       29
#> 31:                                                 Nikkaji       29
#> 32:                                                 Nikkaji       29
#> 33:                                                 Nikkaji       29
#> 34:                                                 Nikkaji       29
#> 35:                                                 Nikkaji       29
#> 36:                                                 Nikkaji       29
#> 37:                                                 Nikkaji       29
#> 38:                                                 Nikkaji       29
#> 39:                                                 Nikkaji       29
#> 40:                                               BindingDB       31
#> 41:                                               BindingDB       31
#> 42: EPA (Environmental Protection Agency) CompTox Dashboard       32
#> 43:                                               LipidMaps       33
#> 44:                                             DrugCentral       34
#> 45:                                            Metabolights       36
#> 46:                                                  Brenda       37
#> 47:                                                  Brenda       37
#> 48:                                                  Brenda       37
#> 49:                                                  Brenda       37
#> 50:                                                  Brenda       37
#> 51:                                                  Brenda       37
#> 52:                                                  Brenda       37
#> 53:                                            ChemicalBook       39
#> 54:                                                DailyMed       45
#> 55:                                          clinicaltrials       46
#> 56:                                          clinicaltrials       46
#> 57:                                          clinicaltrials       46
#> 58:                                          clinicaltrials       46
#> 59:                                          clinicaltrials       46
#> 60:                                          clinicaltrials       46
#> 61:                                          clinicaltrials       46
#> 62:                                          clinicaltrials       46
#> 63:                                          clinicaltrials       46
#> 64:                                          clinicaltrials       46
#> 65:                                          clinicaltrials       46
#> 66:                                                  rxnorm       47
#> 67:                                                  rxnorm       47
#> 68:                                                  rxnorm       47
#> 69:                                                  rxnorm       47
#> 70:                                                  rxnorm       47
#> 71:                                                  rxnorm       47
#> 72:                                                  rxnorm       47
#> 73:                                                  rxnorm       47
#> 74:                                          MedChemExpress       48
#> 75:                                        Probes And Drugs       49
#>                                                    NameLong sourceID
#>                                                                                                                           sourceURL
#>                                                                                                                              <char>
#>  1:                                                                        https://www.ebi.ac.uk/chembldb/compound/inspect/CHEMBL38
#>  2:                                                                                            http://www.drugbank.ca/drugs/DB00755
#>  3:                                                                http://www.ebi.ac.uk/pdbe-srv/pdbechem/chemicalCompound/show/REA
#>  4:                                                      http://www.guidetopharmacology.org/GRAC/LigandDisplayForward?ligandId=2644
#>  5:                                                                              http://pubchem.ncbi.nlm.nih.gov/substance/12014646
#>  6:                                                                                  http://www.genome.jp/dbget-bin/www_bget?C00777
#>  7:                                                                    http://www.ebi.ac.uk/chebi/searchId.do?chebiId=CHEBI%3A15367
#>  8:                                                                                                                    SAM002264647
#>  9:                                                                           http://zinc15.docking.org/substances/ZINC000012358651
#> 10:                                                                             https://www.emolecules.com/cgi-bin/more?vid=1934590
#> 11:                                                                         http://www.ebi.ac.uk/gxa/query?conditionQuery=tretinoin
#> 12:                                                                     http://www.ebi.ac.uk/gxa/query?conditionQuery=retinoic acid
#> 13:                                                           http://www.ebi.ac.uk/gxa/query?conditionQuery=all-trans retinoic acid
#> 14:                                                               http://www.ebi.ac.uk/gxa/query?conditionQuery=9-cis retinoic-acid
#> 15:                                                                        https://precision.fda.gov/uniisearch/srs/unii/5688UTC01R
#> 16:                                                                             https://www.surechembl.org/chemical/SCHEMBL19091395
#> 17:                                                                                 https://www.surechembl.org/chemical/SCHEMBL3145
#> 18:                                                                                       https://www.pharmgkb.org/drug/PA164746900
#> 19:                                                                                      http://www.hmdb.ca/metabolites/HMDB0001852
#> 20:                                                                     http://www.selleckchem.com/products/Tretinoin(Aberela).html
#> 21:                                                                              http://pubchem.ncbi.nlm.nih.gov/substance/14849563
#> 22:                                                                              http://pubchem.ncbi.nlm.nih.gov/substance/14825303
#> 23:                                                                                 http://pubchem.ncbi.nlm.nih.gov/compound/444795
#> 24:                                                                                         http://nmrshiftdb.org/molecule/60021090
#> 25:                                                                            http://identifiers.org/lincs.smallmolecule/LSM-42854
#> 26:                                                                       http://actor.epa.gov/actor/chemical.xhtml?casrn=4759-48-2
#> 27:                                                                        http://actor.epa.gov/actor/chemical.xhtml?casrn=302-79-4
#> 28:                                                                     http://actor.epa.gov/actor/chemical.xhtml?casrn=187175-63-9
#> 29:                                                                  https://www.molport.com/shop/molecule-link/MolPort-000-883-857
#> 30:                                                                     http://jglobal.jst.go.jp/en/redirect?Nikkaji_No=J2.378.058E
#> 31:                                                                     http://jglobal.jst.go.jp/en/redirect?Nikkaji_No=J1.313.469C
#> 32:                                                                       http://jglobal.jst.go.jp/en/redirect?Nikkaji_No=J623.910B
#> 33:                                                                       http://jglobal.jst.go.jp/en/redirect?Nikkaji_No=J970.183D
#> 34:                                                                         http://jglobal.jst.go.jp/en/redirect?Nikkaji_No=J1.518K
#> 35:                                                                       http://jglobal.jst.go.jp/en/redirect?Nikkaji_No=J494.243D
#> 36:                                                                       http://jglobal.jst.go.jp/en/redirect?Nikkaji_No=J690.379G
#> 37:                                                                       http://jglobal.jst.go.jp/en/redirect?Nikkaji_No=J646.157C
#> 38:                                                                       http://jglobal.jst.go.jp/en/redirect?Nikkaji_No=J646.158A
#> 39:                                                                       http://jglobal.jst.go.jp/en/redirect?Nikkaji_No=J528.606I
#> 40:                                               http://www.bindingdb.org/bind/chemsearch/marvin/MolStructure.jsp?monomerid=323588
#> 41:                                                http://www.bindingdb.org/bind/chemsearch/marvin/MolStructure.jsp?monomerid=31883
#> 42:                                                                                 https://comptox.epa.gov/dashboard/DTXSID7021239
#> 43:                                                                  http://www.lipidmaps.org/data/LMSDRecord.php?LMID=LMPR01090019
#> 44:                                                                                            http://drugcentral.org/drugcard/2722
#> 45:                                                                                    http://www.ebi.ac.uk/metabolights/MTBLC15367
#> 46:                                                                 https://www.brenda-enzymes.org/ligand.php?brenda_ligand_id=6481
#> 47:                                                                https://www.brenda-enzymes.org/ligand.php?brenda_ligand_id=12679
#> 48:                                                                https://www.brenda-enzymes.org/ligand.php?brenda_ligand_id=21231
#> 49:                                                                 https://www.brenda-enzymes.org/ligand.php?brenda_ligand_id=2354
#> 50:                                                                 https://www.brenda-enzymes.org/ligand.php?brenda_ligand_id=5126
#> 51:                                                               https://www.brenda-enzymes.org/ligand.php?brenda_ligand_id=193381
#> 52:                                                               https://www.brenda-enzymes.org/ligand.php?brenda_ligand_id=260365
#> 53:                                                               https://www.chemicalbook.com/ChemicalProductProperty_EN_CB6222631
#> 54: https://dailymed.nlm.nih.gov/dailymed/search.cfm?adv=1&labeltype=human&query=ACTIVEMOIETY:(TRETINOIN)+OR+INGREDIENT:(TRETINOIN)
#> 55:                                                                    https://www.clinicaltrials.gov/ct2/results?cond=&term=EUDYNA
#> 56:                                                                 https://www.clinicaltrials.gov/ct2/results?cond=&term=TRETINOIN
#> 57:                                                                    https://www.clinicaltrials.gov/ct2/results?cond=&term=ABEREL
#> 58:                                                                    https://www.clinicaltrials.gov/ct2/results?cond=&term=RENOVA
#> 59:                                                                     https://www.clinicaltrials.gov/ct2/results?cond=&term=AVITA
#> 60:                                                                https://www.clinicaltrials.gov/ct2/results?cond=&term=NSC-122758
#> 61:                                                                  https://www.clinicaltrials.gov/ct2/results?cond=&term=VESANOID
#> 62:                                                                   https://www.clinicaltrials.gov/ct2/results?cond=&term=RETIN A
#> 63:                                                                   https://www.clinicaltrials.gov/ct2/results?cond=&term=RETIN-A
#> 64:                                                             https://www.clinicaltrials.gov/ct2/results?cond=&term=RETINOIC ACID
#> 65:                                                                   https://www.clinicaltrials.gov/ct2/results?cond=&term=ALTRENO
#> 66:                                                           https://mor.nlm.nih.gov/RxNav/search?searchBy=RXCUI&searchTerm=846950
#> 67:                                                           https://mor.nlm.nih.gov/RxNav/search?searchBy=RXCUI&searchTerm=215492
#> 68:                                                           https://mor.nlm.nih.gov/RxNav/search?searchBy=RXCUI&searchTerm=221175
#> 69:                                                           https://mor.nlm.nih.gov/RxNav/search?searchBy=RXCUI&searchTerm=153101
#> 70:                                                            https://mor.nlm.nih.gov/RxNav/search?searchBy=RXCUI&searchTerm=10753
#> 71:                                                           https://mor.nlm.nih.gov/RxNav/search?searchBy=RXCUI&searchTerm=214998
#> 72:                                                           https://mor.nlm.nih.gov/RxNav/search?searchBy=RXCUI&searchTerm=728470
#> 73:                                                          https://mor.nlm.nih.gov/RxNav/search?searchBy=RXCUI&searchTerm=2055004
#> 74:                                                                                         https://www.medchemexpress.com/HY-14649
#> 75:                                                                                  https://www.probes-drugs.org/compound/PD001430
#>                                                                                                                           sourceURL
#> 
#> $UniChem_Mappings
#> $UniChem_Mappings$UniChem.UCI
#> [1] 538323
#> 
#> $UniChem_Mappings$UniChem.InchiKey
#> [1] "SHGAZHPCJJPHSC-YCNIQYBTSA-N"
#> 
#> $UniChem_Mappings$UniChem.Inchi
#> [1] "InChI=1S/C20H28O2/c1-15(8-6-9-16(2)14-19(21)22)11-12-18-17(3)10-7-13-20(18,4)5/h6,8-9,11-12,14H,7,10,13H2,1-5H3,(H,21,22)/b9-6+,12-11+,15-8+,16-14+"
#> 
#> $UniChem_Mappings$UniChem.formula
#> [1] "C20H28O2"
#> 
#> $UniChem_Mappings$UniChem.connections
#> [1] "1-15(8-6-9-16(2)14-19(21)22)11-12-18-17(3)10-7-13-20(18,4)5"
#> 
#> $UniChem_Mappings$UniChem.hAtoms
#> [1] "6,8-9,11-12,14H,7,10,13H2,1-5H3,(H,21,22)"
#> 
#>